requirements.txt
django
djangorestframework
gunicorn
whitenoise[brotil]
django-cors-headers